+++ R/utils_assertive.R
@@ -47,7 +47,7 @@ assert_are_same_length = function(x, y)
 assert_all_are_existing_files = function(x)
 {
   x. <- lazyeval::expr_text(x)
-  if (!all(file.exists(x)))
+  if (!all(file_exists(x)))
     stop(glue::glue("File does not exist."), call. = FALSE)
 }